fix(testflight): robust secret decoding, matching icon, 50 coins per life (#2)
* fix(testflight.yml): tolerate wrapped base64 secrets and name the bad one

* feat(icon): render the app icon from the game's sprite art, 50 coins per life

- scripts/icon compiles the rendering sources so the icon always matches the ninja in the game
- PixelTextNode moved to its own file so the font stays free of SpriteKit

* build(project.yml): recommended xcode 27 settings

* fix(testflight.yml): fail on a bad base64 decode instead of masking it

* fix(audio): play sound effects even when the phone is in silent mode

* feat(backgrounds): repaint parallax layers at sprite density with shading and outlines

* fix(hud): show rex hearts only once the boss wakes up

* feat(blocks): fewer shuriken scrolls, 25 percent when small and one block less in 1-1

* fix(tiles): drop the highlight stripe on the question mark

* feat(art): scrolls replace coins, dinosaurs a quarter larger, rex on the icon

- collectible is a golden scroll with a red ribbon; hud icon and map text follow
- enemy sprites and hitboxes scaled by 1.25 through a fractional sprite scale

* feat(rex): redraw the boss as a real tyrannosaur, walkers avoid lava and spikes

- enemies look one step ahead and turn away from hazard tiles, including rolling Anky balls
- icon shows the new Rex towering over the ninja

* feat(rex): jaw slightly open so the teeth show

+++ Sources/Ninjasaurus/Rendering/PixelSprite.swift
@@ -9,9 +9,9 @@ struct PixelSprite: Equatable, Sendable {
     let height: Int
     let pixels: [PixelColor]
     /// Sprite pixels per game unit.
-    let scale: Int
+    let scale: Double
 
-    init(width: Int, height: Int, pixels: [PixelColor], scale: Int = 1) {
+    init(width: Int, height: Int, pixels: [PixelColor], scale: Double = 1) {
         precondition(pixels.count == width * height, "pixel count mismatch")
         self.width = width
         self.height = height
@@ -19,8 +19,8 @@ struct PixelSprite: Equatable, Sendable {
         self.scale = scale
     }
 
-    var unitWidth: Double { Double(width) / Double(scale) }
-    var unitHeight: Double { Double(height) / Double(scale) }
+    var unitWidth: Double { Double(width) / scale }
+    var unitHeight: Double { Double(height) / scale }
 
     init(palette: Palette, rows: [String]) {
         do {
